проверка отписки от самого себя

+++ yatube/posts/views.py
@@ -129,6 +129,7 @@ def profile_follow(request, username):
 
 @login_required
 def profile_unfollow(request, username):
-    get_object_or_404(Follow, user=request.user,
-                      author__username=username).delete()
+    if username != request.user.username:
+        get_object_or_404(Follow, user=request.user,
+                          author__username=username).delete()
     return redirect('posts:profile', username=username)
